Split EquipmentSet into Equip.Set, Equip.SetSources.GlamourDresser and Structs.FFXIV.MiragePrismMiragePlate
