+++ Ktisis/Interface/Windows/ConfigGui.cs
@@ -8,6 +8,8 @@ using Dalamud.Interface;
 using Ktisis.Util;
 using Ktisis.Localization;
 using Ktisis.Structs.Bones;
+using Ktisis.Structs.Actor.Equip;
+using Ktisis.Structs.Actor.Equip.SetSources;
 
 namespace Ktisis.Interface.Windows {
 	internal static class ConfigGui {
@@ -182,19 +184,19 @@ namespace Ktisis.Interface.Windows {
 		public static void DrawDataTab(Configuration cfg)
 		{
 			ImGui.Spacing();
-			var validGlamPlatesFound = Structs.Actor.EquipmentSetSources.GlamourDresser.CountValid();
+			var validGlamPlatesFound = GlamourDresser.CountValid();
 			GuiHelpers.TextTooltip($"Glamour Plates in memory: {validGlamPlatesFound}  ", $"Found {validGlamPlatesFound} valid Glamour Plates");
 			ImGui.SameLine();
 
 			if (GuiHelpers.IconButtonTooltip(FontAwesomeIcon.Sync, "Refresh Glamour Plate memory for the Sets lookups.\nThis memory is kept after a restart.\n\nRequirements:\n One of these windows must be opened: \"Glamour Plate Creation\" (by the Glamour Dresser) or \"Plate Selection\" (by the Glamour Plate skill)."))
-				Structs.Actor.EquipmentSetSources.GlamourDresser.PopulatePlatesData();
+				GlamourDresser.PopulatePlatesData();
 
 			Components.Equipment.CreateGlamourQuestionPopup();
 
 			ImGui.SameLine();
 			if (GuiHelpers.IconButtonTooltip(FontAwesomeIcon.Trash, "Dispose of the Glamour Plates memory and remove configurations for ALL characters."))
 			{
-				Structs.Actor.EquipmentSets.Dispose();
+				EquipmentSets.Dispose();
 				cfg.GlamourPlateData = null;
 			}
