+++ Ktisis/Structs/Actor/Equip/SetSources/GlamourDresser.cs
@@ -1,0 +1,165 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+using Dalamud.Logging;
+using FFXIVClientStructs.FFXIV.Client.Game;
+using Newtonsoft.Json;
+
+using Ktisis.GameData.Excel;
+using Ktisis.Structs.FFXIV;
+
+
+namespace Ktisis.Structs.Actor.Equip.SetSources
+{
+	public class GlamourDresser
+	{
+
+		// Change this constant if Glamour Plates Number is changed in the Game.
+		// TODO: find this variable from the game's memory
+		public const int _platesNumber = 20;
+
+		public static GlamourPlate[]? Plates = null;
+
+		public static int CountValid()
+		{
+			if (Plates == null) return 0;
+			if (Plates.Length == 0) return 0;
+
+			return Plates.Count((p) => p.IsValid())!;
+		}
+
+		public static void PopulatePlatesData()
+		{
+			var local_character_id = Dalamud.ClientState.LocalContentId;
+			if (local_character_id == 0) return;
+
+			var platesBefore = Plates;
+			GetDataFromDresser(); // if the plate window is open, read it
+			if (Plates == platesBefore) PopupOfferOpenGlamourPlates_open(); // if unsuccessfull, proceed with active dialog
+			else SavePlatesToConfig();
+		}
+		private static bool SavePlatesToConfig()
+		{
+			if (Plates == null) return false;
+
+			if (Ktisis.Configuration.GlamourPlateData == null)
+				Ktisis.Configuration.GlamourPlateData = new Dictionary<string, GlamourPlate[]?>();
+			if (Ktisis.Configuration.GlamourPlateData == null) return false;
+
+			var character_key = $"FFXIV_CHR{Dalamud.ClientState.LocalContentId:X16}";
+			Ktisis.Configuration.GlamourPlateData.TryGetValue(character_key, out var platesFromConfig);
+			if (Plates == platesFromConfig) return false;
+
+			Ktisis.Configuration.GlamourPlateData[character_key] = Plates; // actually save
+
+			bool success = Ktisis.Configuration.GlamourPlateData[character_key] == Plates;
+			if (success) PluginLog.Verbose($"Saved {Ktisis.Configuration.GlamourPlateData[character_key]!.Count((p) => p.IsValid())} Plates for {character_key} into config.");
+			return true;
+		}
+		public static void LoadGlamourPlatesIntoMemory()
+		{
+			var local_character_id = Dalamud.ClientState.LocalContentId;
+			if (local_character_id == 0) return;
+
+			if (Ktisis.Configuration.GlamourPlateData == null) return;
+			var character_key = $"FFXIV_CHR{local_character_id:X16}";
+			Ktisis.Configuration.GlamourPlateData!.TryGetValue(character_key, out Plates);
+			PluginLog.Verbose($"Loaded {CountValid()} Plates for {character_key} into memory.");
+		}
+
+		internal static void PopupOfferOpenGlamourPlates_open()
+		{
+			// This is a way to actively get the data, with the user's authorization
+			if (GameMain.IsInSanctuary())
+				Interface.Components.Equipment.OpenGlamourQuestionPopup();
+		}
+		internal static unsafe void PopupOfferOpenGlamourPlates_confirmed()
+		{
+			var agent = MiragePrismMiragePlates.MiragePlateAgent();
+			agent->Show();
+		}
+
+		public static EquipSlot GlamourPlateSlotToEquipSlot(GlamourPlateSlot slot) => (EquipSlot)((int)slot + ((int)slot > 4 ? 1 : 0));
+
+		internal static unsafe void GetDataFromDresser()
+		{
+			var agent = MiragePrismMiragePlates.MiragePlateAgent();
+			if (agent == null) return;
+			var miragePlates = (MiragePrismMiragePlates*)agent;
+			if (!miragePlates->AgentInterface.IsAgentActive()) return;
+
+			var platePages = miragePlates->Pages;
+			Plates = new GlamourPlate[_platesNumber];
+
+			for (int plateNumber = 0; plateNumber < Plates.Length; ++plateNumber)
+				Plates[plateNumber] = new GlamourPlate(platePages[plateNumber]);
+		}
+
+		[Serializable]
+		public class GlamourPlate
+		{
+			public GlamourPlateItem[] Items = new GlamourPlateItem[12];
+			public bool IsValid() => Items.Any((i) => i.IsValid());
+
+			[JsonConstructor]
+			public GlamourPlate(GlamourPlateItem[] items)
+			{
+				Items = items;
+			}
+
+			internal GlamourPlate(Plate plate)
+			{
+				var fields = typeof(Plate).GetFields();
+				for (int slot = 0; slot < fields.Length; slot++)
+				{
+					MirageItem item = (MirageItem)fields[slot].GetValue(plate)!;
+					Items[slot] = new GlamourPlateItem(item, (GlamourPlateSlot)slot);
+				}
+			}
+
+		}
+
+		[Serializable]
+		public class GlamourPlateItem
+		{
+			public uint ItemId { get; set; }
+			public byte DyeId { get; set; }
+			public GlamourPlateSlot Slot { get; set; }
+			public bool IsValid() => ItemId != 0;
+
+			[JsonConstructor]
+			public GlamourPlateItem(uint itemId, byte dyeId, GlamourPlateSlot slot)
+			{
+				ItemId = itemId;
+				DyeId = dyeId;
+				Slot = slot;
+			}
+
+			internal GlamourPlateItem(MirageItem item, GlamourPlateSlot slot)
+			{
+				ItemId = item.ItemId;
+				DyeId = item.DyeId;
+				Slot = slot;
+			}
+
+		}
+
+		public enum GlamourPlateSlot : uint
+		{
+			MainHand = 0,
+			OffHand = 1,
+			Head = 2,
+			Body = 3,
+			Hands = 4,
+			Legs = 5,
+			Feet = 6,
+			Ears = 7,
+			Neck = 8,
+			Wrists = 9,
+			RightRing = 10,
+			LeftRing = 11,
+		}
+	}
+
+}
